Rétablir Gemma-3-12B-it comme cible principale, corriger bugs découverts par exécution réelle
L'expérimentation de bout en bout a été menée sur Gemma-3-270M-it (validation locale,
6 Go VRAM) avant transfert sur la machine de calcul cible avec Gemma-3-12B-it. Cette
validation a mis au jour et corrigé plusieurs bugs réels :

- src/config.py redevient la source unique de vérité pour le choix modèle/SAE (12b par
  défaut, largeur SAE 16k au lieu de 262k pour une bien meilleure couverture Neuronpedia,
  ~10k/262144 labels observés vs largement plus dense en proportion sur 16k) ; RELEASE_ID
  corrigé (repos HF réels sans suffixe "-res" hérité de GemmaScope v1, vérifié via API).
- neuronpedia_labels.py : ancienne route REST cassée remplacée par le téléchargement
  direct des lots S3 publics (validé : 64377 labels récupérés sur gemma-3-270m-it/65k).
- Bug de dtype critique : DTYPE=bf16 par défaut partout (fp16 overflow silencieusement
  sur les activations massives du residual stream Gemma-3, Loss=nan observé) ; cast
  module-wide erroné de ExtendedSAE retiré (la branche "extra" doit rester fp32).
- gemma_scope_loader.py : hook_layer dynamique, support config.json (renommé depuis
  cfg.json par GemmaScope 2), formats de hook_name multiples.
- download_sae.py réécrit (paramétré via config, télécharge modèle + SAE réels).
- Gardes de robustesse (UMAP, corpus_diff_stats) sur petits corpus.
- Nettoyage : batch_topk_encode mort supprimé, imports cassés corrigés
  (sae_shared.py, compare/pipeline.py), org F2LLM corrigée.
- scripts/retrieval_demo.py : nouveau, câble le module retrieval orphelin.
- README.md et Context.md réécrits pour refléter l'état réel du dépôt.

pytest tests/ : 8/8 passants.

+++ src/sae/frozen_core.py
@@ -72,7 +72,7 @@ class FrozenCoreResidualSAE(nn.Module):
         pre_dead = pre.masked_fill(~dead.unsqueeze(0), float("-inf"))
         vals, idx = pre_dead.topk(k_aux, dim=-1)
         f_aux = torch.zeros_like(pre).scatter_(-1, idx, vals.clamp(min=0.0))
-        e_hat = (f_aux @ self.W_dec_extra) * self.input_scale
+        e_hat = (f_aux @ self.W_dec_extra.float()) * self.input_scale
         return F.mse_loss(e_hat, err) / (err.pow(2).mean() + 1e-8)
 
     def forward(self, x: torch.Tensor) -> dict:
@@ -81,10 +81,17 @@ class FrozenCoreResidualSAE(nn.Module):
             core_acts = self.core_sae.encode(x_bf16)
             core_out = self.core_sae.decode(core_acts)
 
-        residual = x_bf16 - core_out
+        # .float() immédiat : core_sae travaille en bf16 (cf. commentaire decode()) mais
+        # toute la branche "extra" (paramètres, loss) est fp32 partout ailleurs dans ce
+        # fichier. Laisser `residual` en bf16 jusqu'ici et compter sur la promotion
+        # implicite fp32/bf16 dans mse_loss/var_residual/_aux_loss casse le backward
+        # ("Found dtype BFloat16 but expected Float" — la promotion marche en forward
+        # mais pas de façon fiable pour le gradient d'une op mêlant un tenseur fp32
+        # avec grad_fn et un tenseur bf16 sans grad_fn). Confirmé par isolation empirique.
+        residual = (x_bf16 - core_out).float()
         pre = self._pre_extra(residual)
         extra_acts = self.topk_extra(pre)
-        extra_out = (extra_acts @ self.W_dec_extra) * self.input_scale
+        extra_out = (extra_acts @ self.W_dec_extra.float()) * self.input_scale
 
         mse_loss = F.mse_loss(extra_out, residual)
         var_residual = (residual - residual.mean(dim=0)).pow(2).mean()
